Rename get/set to download/upload
