+++ src/templates/default/_dockerignore
@@ -1,5 +1,4 @@
 docker-compose.yml
 .dockerignore
 Dockerfile
-.gitignore
-.git
+.git*
