+++ src/templates/default/_env
@@ -1,2 +1,2 @@
-# Used locally by "yarn set-env" and "yarn get-env" for getting and setting secrets through kubectl
+# Used locally by "yarn upload-env" and "yarn download-env" for getting and setting secrets through kubectl
 GITLAB_PERSONAL_ACCESS_TOKEN={{token}}
